requirements.txt
absl-py==2.1.0
altair==5.3.0
app==0.0.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
docutils==0.21.2
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.63.0
gunicorn==23.0.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
keyring==25.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
namex==0.0.8
nh3==0.2.17
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.2
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
readme_renderer==43.0
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
setuptools==69.5.1
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
streamlit==1.34.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
twine==5.1.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
ushlex==0.99.1
virtualenv==20.26.2
watchdog==4.0.0
websocket-client==1.8.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
zipp==3.18.2
